Big changes with the code structure. Add cogs to separate commands and main file. The bot usage is the same, but the structure is easier to use for the coder. Cogs are stored in cogs. Add a file `utils.py` with functions that takes too many places.
