+++ main.py
@@ -1,46 +1,20 @@
 import discord
 from discord.ext import commands
+from cogs.set_channel import SetChannel
+from cogs.config_channel import ConfigChannel
 import json
 import os
 import sys
-import logging
+from utils import update_json, update_data, create_logger, help_embed
 
 
 PATH = os.path.dirname(__file__)
-DATA_PATH = os.path.join(PATH, "data.json")
 CONFIG_PATH = os.path.join(PATH, "config.json")
-LOG_PATH = os.path.join(PATH, ".logs")
 
 data = {}
-bot = commands.Bot(command_prefix="$", help_command=None)
+bot = commands.Bot(command_prefix="!", help_command=None)
 
-logger = logging.getLogger("pvc bot")
-logger.setLevel(logging.INFO)
-file_handler = logging.FileHandler(LOG_PATH, "a")
-file_handler.setFormatter(logging.Formatter(r"%(asctime)s - [%(levelname)s] ~ %(message)s"))
-stream_handler = logging.StreamHandler(sys.stdout)
-stream_handler.setFormatter(logging.Formatter(r"[%(levelname)s] ~ %(message)s"))
-logger.addHandler(file_handler)
-logger.addHandler(stream_handler)
-
-
-def format_time(t: int):
-    minutes, seconds = divmod(t, 60)
-    text = str(seconds)
-    if minutes:
-        text = f"{minutes} minutes and {seconds}"
-    return f"You are going too fast ! Please wait {text} seconds before retry it."
-
-
-def update_json(data: dict):
-    with open(DATA_PATH, "w") as file:
-        json.dump(data, file, indent=4)
-
-
-def update_data(data: dict):
-    with open(DATA_PATH, "r") as file:
-        for key, value in json.load(file).items():
-            data[key] = value
+logger = create_logger()
 
 
 def get_token():
@@ -99,270 +73,28 @@ async def on_command_error(ctx, error):
 
 @bot.command(name="help")
 async def _help(ctx):
-    embed = discord.Embed(
-        title="Help",
-        color=discord.Color(5027441)
-    )
-    embed.set_thumbnail(url=f"https://cdn.discordapp.com/avatars/{bot.user.id}/{bot.user.avatar}.png?size=1024")
-    embed.add_field(
-        name="__help__",
-        value="**Description:** Show you this message.\n**Usage**: `help`",
-        inline=False
-    )
-    embed.add_field(
-        name="__setchannel__",
-        value="**Description:** allows you to assign a main channel.\n**Usage**: `setchannel [channel_id]`",
-        inline=False
-    )
-    embed.add_field(
-        name="__unsetchannel__",
-        value="**Description:** reset the main channel.\n**Usage**: `unsetchannel`",
-        inline=False
-    )
-    embed.add_field(
-        name="__set__",
-        value="**Description:** allows you to modify your channel once inside.\n**Usage**: `set [arg]`",
-        inline=False
-    )
-    await ctx.send(embed=embed)
-
-
-@bot.command()
-@commands.has_permissions(administrator=True)
-async def setchannel(ctx):
-    content = ctx.message.content.split()
-
-    if len(content) <= 1:
-        await ctx.send("Please specify a channel id.")
-        return
-
-    if not content[1].isdigit():
-        await ctx.send("Please use a valid id.")
-        return
-
-    channel = bot.get_channel(int(content[1]))
-
-    if not channel:
-        await ctx.send("Please use a valid id.")
-        return
-
-    if not isinstance(channel, discord.VoiceChannel):
-        await ctx.send("Please use a voice channel id.")
-        return
-
-    if data[str(ctx.guild.id)]["main"] is not None:
-        logger.warning(f"Channel already set {data[str(ctx.guild.id)]['main']}.")
-    logger.info(f"Set main channel <{channel.name}.{channel.id}>")
-
-    data[str(ctx.guild.id)]['main'] = channel.id
-    await ctx.send("Main channel set.")
-
-    update_json(data)
-    update_data(data)
-
-
-@bot.command()
-@commands.has_permissions(administrator=True)
-async def unsetchannel(ctx):
-    data[str(ctx.guild.id)]['main'] = None
-    await ctx.send("Main channel reset.")
-
-    logger.info("Main channel reset.")
-
-    update_json(data)
-    update_data(data)
-
-
-def help_embed(member):
-    embed = discord.Embed(
-        title="Settings help",
-        color=discord.Color(5027441)
-    )
-    embed.set_thumbnail(url=f"https://cdn.discordapp.com/avatars/{member.id}/{member.avatar}.png?size=1024")
-    embed.add_field(
-        name="__owner__",
-        value="**Description:** allows you to change the owner\n"
-              "**Usage**: `set owner [mention]`",
-        inline=False
-    )
-    embed.add_field(
-        name="__name__",
-        value="**Description:** allows you to change the channel's name\n"
-              "**Usage**: `set name [name]`",
-        inline=False
-    )
-    embed.add_field(
-        name="__places__",
-        value="**Description:** allows you to change the number of places in the channel.\n"
-              "Take a number between 0 to 99, 0 mean no limit.\n"
-              "**Usage**: `set places [int]`",
-        inline=False
-    )
-    embed.add_field(
-        name="__public/private__",
-        value="**Description:** allows you to define the channel as public or private\n"
-              "**Usage**: `set private` or `set public`",
-        inline=False
-    )
-    embed.add_field(
-        name="__hide/reveal__",
-        value="**Description:** allows you to define the channel as hidden or visible\n"
-              "**Usage**: `set hide` or `set reveal`",
-        inline=False
-    )
-    embed.add_field(
-        name="__invite/kick__",
-        value="**Description:** allows you to invite or kick somebody from the channel\n"
-              "**Usage**: `set invite [mention]` or `set kick [mention]`",
-        inline=False
-    )
-    embed.add_field(
-        name="__help__",
-        value="**Description:** Show you this embed\n"
-              "**Usage**: `set help`",
-        inline=False
-    )
-    return embed
-
-
-@bot.group(name="set")
-async def _set(ctx):
-    content = ctx.message.content.split()
-    voice = ctx.author.voice
-    member_id = ctx.author.id
-    guild_id = str(ctx.guild.id)
-    keywords = ["private", "public", "owner", "places", "name", "hide", "reveal", "invite", "kick"]
-
-    if not voice:
-        await ctx.send('You must be in your voice channel.')
-        return
-
-    chan = voice.channel
-    voice = str(voice.channel.id)
-
-    if voice not in data[guild_id]['channels'].keys():
-        await ctx.send('You must be in your voice channel.')
-        return
-
-    if member_id != data[guild_id]['channels'][voice]["owner"]:
-        await ctx.send("You don't have permission to do this !")
-        return
-
-    if len(content) <= 1 or content[1] not in keywords:
-        await ctx.send(embed=help_embed(ctx.author))
-        return
-
-# wtf?
-    if content[1] == "name":
-        return
-
-    elif content[1] == "owner" and len(ctx.message.mentions):
-        data[guild_id]['channels'][voice]["owner"] = ctx.message.mentions[0].id
-        await ctx.message.channel.send(f"Owner successfully changed to {ctx.message.mentions[0].name}.")
-
-    elif content[1] == "places":
-        if not content[2].isdigit() or not 0 <= int(content[2]) < 100:
-            await ctx.send("Please use a number between 0 and 99.")
-            return
-
-        places_count = int(content[2])
-
-        await chan.edit(user_limit=places_count)
-        await ctx.send("Number of places successfully changed.")
-        data[guild_id]['channels'][voice]["places"] = places_count
-
-    elif content[1] == "reveal":
-        await chan.set_permissions(ctx.guild.default_role, view_channel=True)
-        await ctx.send("The channel is now visible.")
-        data[guild_id]['channels'][voice]["visible"] = True
-
-    elif content[1] == "hide":
-        await chan.set_permissions(ctx.guild.default_role, view_channel=False)
-        await ctx.send("The channel is now hide.")
-        data[guild_id]['channels'][voice]["visible"] = False
-
-    elif content[1] == "private":
-        await chan.set_permissions(ctx.guild.default_role, connect=False)
-        await ctx.send("Channel defined as private.")
-
-    elif content[1] == "public":
-        await chan.set_permissions(ctx.guild.default_role, connect=True)
-        await ctx.send("Channel defined as public.")
-
-# this is not in the right place
-#    if len(content) < 2 or not len(ctx.message.mentions):
-#        await ctx.send("Please mention a valid user.")
-#        return
-
-    elif content[1] == "invite":
-        await chan.set_permissions(ctx.message.mentions[0], connect=True, view_channel=True)
-        await ctx.send(f"{ctx.message.mentions[0]} is now allowed to connect to the channel.")
-
-    else:
-        member = ctx.message.mentions[0]
-        await chan.set_permissions(member, connect=False)
-        await member.move_to(None)
-        await ctx.send(f"{member} is now forbidden to connect to the channel.")
-
-    update_json(data)
-    update_data(data)
-
-
-@commands.cooldown(1, 300, commands.BucketType.member)
-@_set.command()
-async def name(ctx):
-    content = ctx.message.content.split()
-    voice = ctx.author.voice
-    member_id = ctx.author.id
-    guild_id = str(ctx.guild.id)
-
-    if not voice:
-        return name.reset_cooldown(ctx)
-
-    chan = voice.channel
-    voice = str(voice.channel.id)
-
-    if voice not in data[guild_id]['channels'].keys():
-        return name.reset_cooldown(ctx)
-
-    if member_id != data[guild_id]['channels'][voice]["owner"]:
-        return name.reset_cooldown(ctx)
-
-    if len(content) > 2:
-        await chan.edit(name=" ".join(content[2:]))
-        await ctx.message.channel.send("Name successfully changed.")
-        data[guild_id]['channels'][voice]["name"] = " ".join(content[2:])
-    else:
-        return name.reset_cooldown(ctx)
-
-
-@name.error
-async def on_command_error(ctx, error):
-    logger.error(f"Error occured: {error}")
-    if isinstance(error, commands.CommandOnCooldown):
-        await ctx.send(format_time(round(error.retry_after)))
+    await ctx.send(embed=help_embed)
 
 
 @bot.event
 async def on_voice_state_update(member, before, after):
-    if after.channel and after.channel.id == data[str(member.guild.id)]['main']:  # if the user connects to the set voice channel...
+    # if the user connects to the set voice channel...
+    if after.channel and after.channel.id == data[str(member.guild.id)]['main']:
         category = bot.get_channel(after.channel.category_id)
         # create a new channel, named after the member name, set permissions, and move the member to it
         overwrite = discord.PermissionOverwrite(manage_permissions=True, connect=True,
-                                                manage_channels=True, view_channel=True, move_member=True)
-        change = await member.guild.create_voice_channel(f"{member.name}'s channel", overwrites={bot.user: overwrite}, category=category)
+                                                manage_channels=True, view_channel=True, move_members=True)
+        change = await member.guild.create_voice_channel(f"{member.name}'s channel",
+                                                         overwrites={bot.user: overwrite}, category=category)
         await change.set_permissions(member, view_channel=True, connect=True)
         await member.move_to(change)
 
         logger.info(f"{member.name} joined and was moved to \"{member.name}'s channel\"")
         # update the data
-        data[str(member.guild.id)]["channels"][str(change.id)] = {
-            "owner": member.id,
-            "name": change.name,
-            "public": True,
-            "places": 0
-        }
-    if before.channel and str(before.channel.id) in data[str(member.guild.id)]["channels"]:  # if the user disconnects from a personnal voice channel
+        data[str(member.guild.id)]["channels"][str(change.id)] = member.id
+
+    # if the user disconnects from a personnal voice channel
+    if before.channel and str(before.channel.id) in data[str(member.guild.id)]["channels"]:
         if not len(before.channel.members):  # if there is no more users in the voice channel, delete it
             data[str(member.guild.id)]['channels'].pop(str(before.channel.id))
             await before.channel.delete(reason='Last member leave')
@@ -371,6 +103,9 @@ async def on_voice_state_update(member, before, after):
     update_json(data)
     update_data(data)
 
+
+bot.add_cog(SetChannel(bot, data, logger))
+bot.add_cog(ConfigChannel(bot, data, logger))
 
 try:
     bot.run(get_token())
